rename AbstractBatchEventHandler
